% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Data set}
\format{
A data frame
\describe{
   \item{DDT}{}
   \item{LENGTH}{}
   \item{MILE}{}
   \item{RIVER}{}
   \item{SPECIES}{}
   \item{WEIGHT}{}
}
}
\usage{
ddt
}
\description{
Set of data for MATH4753
}
\keyword{datasets}
