% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Data set}
\format{
A data frame
\describe{
   \item{DISTANCE}{}
   \item{DAMAGE}{}
}
}
\usage{
fire
}
\description{
Set of data for MATH4753
}
\keyword{datasets}
