% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mychisim.R
\name{mychisim}
\alias{mychisim}
\title{chi-square simualtion}
\usage{
mychisim(iter = 1000, n = 10)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Number of sample size}
}
\description{
chi-square simualtion
}
