% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{CLT Poisson}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Sample Size}

\item{iter}{Number of iterations}

\item{lambda}{a vector of non negative means}

\item{...}{Other settings}
}
\value{
A plot
}
\description{
Central limit function using Poisson's regression.
}
\examples{
mycltp(20,1000,10)
}
