% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{myf - " quadratic linear model}
\usage{
myf(x, coef)
}
\arguments{
\item{x}{Set an X intercept}

\item{coef}{Input a list of coefficients of a linear! model}
}
\description{
myf - " quadratic linear model
}
