% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Curve Probability Distribution}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{S.D.}

\item{a}{user input}
}
\value{
A graph representing probabiltiy through area
}
\description{
Normal Curve Probability Distribution
}
\examples{
myncurve(mu=4, sigma=2, a=5)
}
