% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynewt.R
\name{mynewt}
\alias{mynewt}
\title{Newton-Raphson Algorithm}
\usage{
mynewt(x0, delta = 0.001, f, fdash)
}
\arguments{
\item{x0}{X-not first approximation value}

\item{delta}{increment, try 0.001}

\item{f}{Function}

\item{fdash}{Function derivative}
}
\description{
The Newton-Raphson Algorithm, which can be used to
   find the roots of a function.
}
\examples{
mynewt(x0=10,delta=0.000001,f=function(x) x^2-4,fdash=function(x) 2*x )
}
