% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynewt2.R
\name{mynewt2}
\alias{mynewt2}
\title{Newton-Raphson Algorithm Self Deriving}
\usage{
mynewt2(x0, delta = 0.001, f, h = 1e-04)
}
\arguments{
\item{x0}{X-not first approximation value}

\item{delta}{increment, try 0.001}

\item{f}{Function}

\item{h}{derivative increment, try 0.0001}
}
\description{
The Newton-Raphson Algorithm, which can be used to
   find the roots of a function. Capable of deriving
   desiredfunction.
}
\examples{
mynewt2(x0=10,delta=0.000001,f=function(x) x^2-4)
}
