% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzbirds.R
\docType{data}
\name{nzbirds}
\alias{nzbirds}
\title{NZBIRDS Data set}
\format{
A data frame
\describe{
   \item{Species}{}
   \item{Name}{}
   \item{Extinct}{}
   \item{Habitat}{}
   \item{Nest.Site}{}
   \item{Nest.Density}{}
   \item{Diet}{}
   \item{Flight}{}
   \item{Body.Mass}{}
   \item{Egg.Length}{}
}
}
\usage{
nzbirds
}
\description{
Set of data for MATH4753
}
\keyword{datasets}
