% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.r
\name{pareto}
\alias{pareto}
\title{Pareto Barplot}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{Input}

\item{mn}{Main Title}

\item{...}{Graph parameters}
}
\description{
Pareto Barplot
}
