% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Scatter Histogram}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{A vector of independant variables}

\item{y}{A vector of dependant variables}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}
}
\description{
Creates a Scatter Histogram from a given set of X-Y data.
}
