% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.score.R
\name{z.score}
\alias{z.score}
\title{Z-score}
\usage{
z.score(x)
}
\arguments{
\item{x}{a quantitative vector}
}
\value{
Z-score for a data set, the squared values
}
\description{
Z-score
}
\examples{
z.score(x=c(1,2,3,1,2,3,1,2,3,6,6,8,9))
}
